requirements.txt
Django>=3.2,<3.3
wagtai==2.14.1
